requirements.txt
# Voice RAG Chatbot Dependencies

# Core
python-dotenv>=1.0.0
pydantic>=2.7.4
pydantic-settings>=2.1.0

# Gradio UI
gradio>=4.44.0

# STT - Whisper
faster-whisper>=1.0.0

# TTS
gTTS>=2.5.0

# Emotion Recognition
librosa>=0.10.1
numpy>=1.26.0
torch>=2.1.0
funasr>=1.0.0

# RAG - LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.4
langgraph>=0.2.0

# Vector Store
chromadb>=0.4.22

# Embeddings
sentence-transformers>=2.2.2

# Audio Processing
soundfile>=0.12.1
scipy>=1.11.0

# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
aiofiles>=23.2.1

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
